Update the workflows to the latest versions
Notable here is that we do update the setup scripts to the latest versions which means that we can remove the specific workarounds.
There is a bug in that version that causes Windows builds to fail, but since we do not support windows, we can safely use it here.
This also updates staticcheck to 0.2.0 (2021.1), see https://staticcheck.io/changes/2021.1 for changelog.